+++ pointers/src/main/kotlin/de/md5lukas/waypoints/pointers/variants/PointerVariant.kt
@@ -4,7 +4,6 @@ import de.md5lukas.schedulers.AbstractScheduler
 import de.md5lukas.waypoints.pointers.Pointer
 import de.md5lukas.waypoints.pointers.PointerManager
 import de.md5lukas.waypoints.pointers.config.PointerConfiguration
-import org.bukkit.Bukkit
 import org.bukkit.entity.Player
 
 enum class PointerVariant(
@@ -17,11 +16,7 @@ enum class PointerVariant(
   BLINKING_BLOCK("blinkingBlock", { it.blinkingBlock.enabled }, ::BlinkingBlockPointer),
   BOSS_BAR("bossBar", { it.bossBar.enabled }, ::BossBarPointer),
   COMPASS("compass", { it.compass.enabled }, ::CompassPointer),
-  HOLOGRAM(
-      "hologram",
-      { it.hologram.enabled && Bukkit.getPluginManager().isPluginEnabled("ProtocolLib") },
-      ::HologramPointer,
-  ),
+  HOLOGRAM("hologram", { it.hologram.enabled }, ::HologramPointer),
   PARTICLE("particle", { it.particle.enabled }, ::ParticlePointer),
   TRAIL("trail", { it.trail.enabled }, ::TrailPointer)
 }